[FIX] move BigTitle component to atom
